Add compulsory type conversion to bin_power
Hi professor Bao! 
The  pyeeg reported "TypeError: slice indices must be integers or None or have an __index__ method" when I was using the  bin_power() function with  py3.6. I also tried it in py2.7 but the same error was reported. Then I found that with compulsory type conversion added to the parameters :numpy.floor(Freq / Fs * len(X)) and numpy.floor(Next_Freq / Fs * len(X)), it finally works.
So, I think this change many be safier and less confused to the other users.
Thank you for your work and your contribution to this program!!! 
过年好！